feat(standard_layout): Move auth from internal/ to pkg/
